+++ src-bare/app/app.component.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+
+@Component({
+    moduleId: module.id,
+    selector: 'cmp-prefix1-app-root',
+    templateUrl: 'app.component.html',
+    styleUrls: ['app.component.css']
+})
+
+
+export class AppComponent {
+
+    constructor() {}
+
+}