+++ src-bare/public/system.config.prod.js
@@ -1,0 +1,15 @@
+(function (global) {
+  System.defaultJSExtensions = true;
+  System.config({
+    'map': {
+      'bundle': 'bundle.js',
+      'vendor': 'vendor.js'
+    },
+    'meta': {
+      'bundle': {
+        deps: []
+      }
+    }
+  });
+})(this);
+