+++ src/app/app.component.e2e-spec.ts
@@ -1,6 +1,0 @@
-import { browser, element, by } from 'protractor';
-
-describe('angular2-rollup E2E Tests', function () {
-
-
-});